backend/pyproject.toml
[tool.poetry]
name = "neena-core"
version = "0.1.0"
description = "Neena Core project"
authors = ["Rauf <rauf@neena.io>"]

[tool.poetry.dependencies]
python = "^3.9.4"
inboard = "^0.37.0"
python-multipart = "^0.0.5"
fastapi = "^0.89.0"
requests = "^2.28.1"
celery = "^5.2.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.1.0"
raven = "^6.10.0"
alembic = "^1.8.1"
email-validator = "^1.3.0"
sqlalchemy = "^2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.23.1"
psycopg2-binary = "^2.9.5"
setuptools = "^65.6.3"
argon2-cffi = "^21.3.0"
attrs = "^22.0.0"
opencensus-ext-azure = "^1.1.13"
fastapi-auth0 = "^0.5.0"


[tool.poetry.dev-dependencies]
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.11.2"
autoflake = "^2.0.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"
sqlalchemy2-stubs = "^0.0.2a29"
pytest-cov = "^4.0.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
